changed short param to int in AreaEval.containsColumn() method
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@891425 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/formula/eval/AreaEval.java
@@ -68,7 +68,7 @@ public interface AreaEval extends TwoDEval {
      * returns true if the specified col is in range
      * @param col
      */
-    boolean containsColumn(short col);
+    boolean containsColumn(int col);
 
     /**
      * returns true if the specified row is in range
